Add n-dimensional valueFunc to interpolation.py
